<Fix> weekly메뉴 컴포넌트내 유틸함수 컴포넌트 분리
