+++ src/components/atoms/Meal/Web/WeeklyMenuView.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import type { MenuItem } from '@/api/menu';
 import MenuDayButton from '@/components/molecules/MenuDayButton';
 import MenuItemButton from '@/components/molecules/MenuItemButton';
+import { useMenuData } from '@/hooks/menu/useMenuData';
 
 const timeLabelMap = {
   BREAKFAST: '아침',
@@ -9,37 +10,6 @@ const timeLabelMap = {
   DINNER: '저녁',
 } as const;
 const ORDER: Array<keyof typeof timeLabelMap> = ['BREAKFAST', 'LUNCH', 'DINNER'];
-
-// 주차/기간 유틸
-function startOfWeek(d: Date, weekStartsOn = 1) {
-  const date = new Date(d);
-  const day = date.getDay();
-  const diff = (day - weekStartsOn + 7) % 7;
-  date.setDate(date.getDate() - diff);
-  date.setHours(0, 0, 0, 0);
-  return date;
-}
-function endOfWeek(d: Date, weekStartsOn = 1) {
-  const start = startOfWeek(d, weekStartsOn);
-  const end = new Date(start);
-  end.setDate(start.getDate() + 6);
-  end.setHours(23, 59, 59, 999);
-  return end;
-}
-function getWeekOfMonth(d: Date, weekStartsOn = 1) {
-  const date = new Date(d);
-  const month = date.getMonth();
-  const firstOfMonth = new Date(date.getFullYear(), month, 1);
-  const firstWeekStart = startOfWeek(firstOfMonth, weekStartsOn);
-  const thisWeekStart = startOfWeek(date, weekStartsOn);
-  const diffDays = Math.floor((thisWeekStart.getTime() - firstWeekStart.getTime()) / 86400000);
-  return Math.floor(diffDays / 7) + 1;
-}
-function fmtMD(date: Date) {
-  const m = String(date.getMonth() + 1).padStart(2, '0');
-  const d = String(date.getDate()).padStart(2, '0');
-  return `${m}.${d}`;
-}
 
 type WeeklyMenuViewProps = {
   groupedByDate: Array<[string, MenuItem[]]>;
@@ -54,6 +24,7 @@ export default function WeeklyMenuView({
   nowCategory,
 }: WeeklyMenuViewProps) {
   const todayDate = todayKey;
+  const { startOfWeek, endOfWeek, getWeekOfMonth, fmtMD } = useMenuData();
 
   const now = new Date();
   const weekStartsOn = 1;
